Refactor MetadataProperty into its own files
This class in PntsToGltfConverter.cpp is useful for dealing with
instance metadata too.
